+++ src/lib.rs
@@ -70,6 +70,8 @@
 // std builds.
 extern crate alloc;
 
+#[cfg(feature = "experimental-multiprocess")]
+pub use db::ConcurrencyMode;
 #[cfg(not(redb_no_std))]
 pub use db::ReadOnlyDatabase;
 pub use db::{