+++ src/tree_store/btree_mutator.rs
@@ -1849,6 +1849,7 @@ mod tests {
             None,
             0,
             false,
+            crate::db::ConcurrencyMode::SingleProcess,
         )
         .unwrap();
         mem.reset_allocator_state().unwrap();